requirements.txt
selenium
beautifulsoup4
schedule
python-dotenv
